requirements.txt
allure-pytest==2.8.22
allure-python-commons==2.8.22
pytest==6.1.2
PyYAML==5.3.1
requests==2.25.0
